fixed the makeAdmin

+++ app/Http/Controllers/AdminController.php
@@ -14,8 +14,9 @@ class AdminController extends Controller
             return $this->validateAdmin();
         }
 
-        // return User::paginate(15);
-        return User::all();
+        return User::where('is_admin', '=', 0)
+        // ->paginate(15);
+        ->get();
     }
 
     public function showAdmins()
@@ -93,7 +94,8 @@ class AdminController extends Controller
         if (!Hash::check(request()->admin_password, $authed_user->password)) {
             return response('wrong admin password ', 401);
         }
-        $user->is_admin = true;
+        $user->is_admin = 1;
+        $user->save();
         return [
             'message'=> 'user now is admin',
             'user' => $user
@@ -109,8 +111,12 @@ class AdminController extends Controller
             return $this->validateAdmin();
         }
 
-        $user->delete();
-        return "user deleted successfully";
+        if (!$user->is_admin) {
+            $user->delete();
+            return "user deleted successfully";
+        } else {
+            return response("You can't delete other Admins", 400) ;
+        }
     }
 
     protected function validateAdmin()
